change: removing suggested packages customization for metrumrg and making related changes.
I am adding a script to turn the pkgr-installed packages into a local
repo, which can then be used for deployment. This lets us circumvent SSL
issues that randomly pop up between RSConnect deployments and our
TFLCRAN repo.

Additionally, I have modified the pkgr.yml file to remove Suggested
dependencies for metrumrg, as well as the 'reticulate' package, so that
this application is no longer dependant on reticulate.

Notably, packrat still detects reticulate as a required dependency. As
far as I can tell from looking at the manifest.json, reticulate is ONLY
a SUGGESTED dependency of knitr, so at this point I think this is a bug
in packrat or rsconnect.

To circumvent THIS, I have added the option to ignore reticulate into
the packrat.opts file.